refactor: make dashboard printing code more typed
Extract the core logic for 'print these PR entries' into a separate function,
and use this opportunity to make the printing code more typed.

This will allow manually filtering a list of PRs on the Python side.

I verified locally that this generated virtually identical HTML output:
the only difference are 'last updated on' date and the 'X days ago',
as expected.