change: move the queue view back to queue.json, for now
The metadata-based computation still struggles with keeping everything up to
date, and I have not yet fully identifier or understood why. (Two reasons
are known; it seems that force-pushes are not recognised as 'new data, please
downloaded updated one', and ongoing CI runs finishing are neither.
This list is probably not exhaustive!)
In the meantime, show the correct data, to prevent confusing people.
We leave the comparison output in, so investigation can continue in parallel.

We also disable the diagnostic comparison for the 'needs-merge' dashboard,
since its output so far has been stable, and we can save that one API call
per script run for later.