refactor: hide the 'last updated' field on the assignment dashboard
- Add configuration for hiding the 'last updated' field, to make this generally
available.
- Update the builder pattern methods: still a bit of boilerplate.
Upside: they mean I don't need to remember the argument names (good).
Using named arguments works if you want to specify *all* or *most* args.
For changing just one argument, and using the default elsewhere,
builders are nicer.

- Use this to hide the 'updated' yield on the assignement dashboard.
If this is a good choice can be discussed. Without it, the dashboard looks
bad on my screen (and enabling this improves it). This can be tweaked later!