fix: fix sorting by total review time (slightly hacky)
- pass the full time difference into the column, to allow structured
parsing of the relevant information. Add a new formatdelta2 function,
which prints this information. (Currently, we deal with year and month fields
in a rather hacky way. TODO: fix this in a more principled way, by returning
the relativedelta plus the time in days, and using the latter for this field).
- put this information inside a display:none div, so it is hidden
- make sure "unknown" information also contains this:
then, we can use this to auto-detect a 'total time in this state' column
- implement a custom normaliser, for sorting these columns correctly.
Pay attention to not run into undesirable implicit conversions.

The new sorting seems to be almost what I need.