fix(*.yml): correct 'if' conditions in workflow steps
In publish_dashboard.yml, running them linearly (and aborting if one step
fails) makes perfect sense now: the only expected change is from
updating the webpage, and each step requires the previous ones.

In gather_stats.yml, make sure to download the json files and commit
and push even if a previous step fails: not all of them are essential.
The .json files, however, *are* essential for most steps.