WIP-feat(process.py): add information about the last update to aggregate files
This is advantageous for several reasons:
- currently, several places in the dashboard probe for a PR's events data
They read the data file again, parse the events data and compute the relevant
information. Having this in one place is much clearer data flow.

- efficiency: computing all information in one place means I only have to
compute it once (currently, this can happen several times). It also makes it
easier to cache, and also allows caching the metadata computation for a PR.

- re-usability: other jobs can access this information easily

This commit is not complete yet: it does not parse or expose the aggregate
information on the dashboard, and the .json generation in fact doesn't work
yet.