feat(process.py): drop threshold from generated assignment data
The PR metadata is now complete \o/

Accordingly, the reviewer suggestion page will display all assignments
ever made to this reviewer. That is misleading (as reviewers could have been
active for different time spans) --- but the numbers are skewed even more
by the fact that many reviewers don't use the assignee field much.

In the future, we may consider filtering this to e.g. the last 12 months.
For now, we just add a word of caution about interpreting this.