refactor(suggest_reviewers): extract the review suggestions into a structure
At the end of the day, there are two useful kinds of suggestions.
The first is "everybody whose interests match, even if busy"
(for a webpage for manual assignments; those over capacity could be
styled differently on the page),
the other is "all reviewers on the rotation whose capacity matches".
We will add the latter in the next step"