feat(process.py): add a "fast mode"
When passed the --fast option, do not update the aggregate file of all
pull requests (but only the one for open PRs). This cuts the time to run
this script more than in half.
(Since the backend workflow runs this twice, we can use this for the first
run and speed up the overall workflow a bit.)