change: broaden the definition of stale unassigned PRs
Already list unassigned PRs on the queue with no status change in *one week*.
The number of stale unassigned PRs is low enough now, that we can probably
try this. One week is also long enough that "stumbling on this PR because
it was recent" is unlikely enough.

Moreover, fix comments about the definition of the "stale assigned" queue.
I'll call the current measure good enough. I will have false positives
(if one wishes to call them so) for PRs that were forgotten to label; I'd
say that's a feature.